feat: added basic layout and info user can edit
