+++ src/app/profile/page.tsx
@@ -28,7 +28,7 @@ export default async function ProfilePage() {
 
   return (
     <div className="max-w-2xl mx-auto mt-10 p-6 rounded-md shadow-md bg-white">
-      <h1 className="text-xl font-bold mb-2">{user.name}</h1>
+      <h1 className="text-xl font-bold mb-2 text-dark">{user.name}</h1>
       <p className="text-gray-500">{user.email}</p>
 
       <ProfileEditor user={user} />
@@ -36,20 +36,20 @@ export default async function ProfilePage() {
       {/* Stats */}
       <div className="grid grid-cols-2 gap-4 text-center mt-6">
         <div>
-          <p className="text-2xl font-bold">{user.completedWorkouts}</p>
           <p className="text-sm text-gray-600">Workouts</p>
+          <p className="text-2xl font-bold text-dark">{user.completedWorkouts}</p>
         </div>
         <div>
-          <p className="text-2xl font-bold">{user.completedExercises}</p>
-          <p className="text-sm text-gray-600">Exercises</p>
+        <p className="text-sm text-gray-600">Exercises</p>
+          <p className="text-2xl font-bold text-dark">{user.completedExercises}</p>
         </div>
         <div>
-          <p className="text-2xl font-bold">{user.xp}</p>
-          <p className="text-sm text-gray-600">XP</p>
+        <p className="text-sm text-gray-600">XP</p>
+          <p className="text-2xl font-bold text-dark">{user.xp}</p>
         </div>
         <div>
-          <p className="text-2xl font-bold">{user.level}</p>
-          <p className="text-sm text-gray-600">Level</p>
+        <p className="text-sm text-gray-600">Level</p>
+          <p className="text-2xl font-bold text-dark">{user.level}</p>
         </div>
       </div>
     </div>
